requirements.txt
aiohttp
feedparser
python-dateutil
supabase
